Move types from php_v8js_macros.h to new headers
