+++ v8js_class.h
@@ -14,6 +14,58 @@
 #ifndef V8JS_CLASS_H
 #define V8JS_CLASS_H
 
+
+/* Abbreviate long type names */
+typedef v8::Persistent<v8::FunctionTemplate, v8::CopyablePersistentTraits<v8::FunctionTemplate> > v8js_tmpl_t;
+typedef v8::Persistent<v8::Object, v8::CopyablePersistentTraits<v8::Object> > v8js_persistent_obj_t;
+
+/* Forward declarations */
+struct v8js_v8object;
+struct php_v8js_accessor_ctx;
+
+/* {{{ Context container */
+struct php_v8js_ctx {
+  zend_object std;
+  v8::Persistent<v8::String> object_name;
+  v8::Persistent<v8::Context> context;
+  zend_bool report_uncaught;
+  zval *pending_exception;
+  int in_execution;
+  v8::Isolate *isolate;
+
+  long time_limit;
+  bool time_limit_hit;
+  long memory_limit;
+  bool memory_limit_hit;
+
+  v8::Persistent<v8::FunctionTemplate> global_template;
+
+  zval *module_loader;
+  std::vector<char *> modules_stack;
+  std::vector<char *> modules_base;
+  std::map<char *, v8js_persistent_obj_t> modules_loaded;
+  std::map<const char *,v8js_tmpl_t> template_cache;
+
+  std::map<zval *, v8js_persistent_obj_t> weak_objects;
+  std::map<v8js_tmpl_t *, v8js_persistent_obj_t> weak_closures;
+
+  std::list<v8js_v8object *> v8js_v8objects;
+
+  std::vector<php_v8js_accessor_ctx *> accessor_list;
+  char *tz;
+#ifdef ZTS
+  void ***zts_ctx;
+#endif
+};
+/* }}} */
+
+#ifdef ZTS
+# define V8JS_TSRMLS_FETCH() TSRMLS_FETCH_FROM_CTX(((php_v8js_ctx *) isolate->GetData(0))->zts_ctx);
+#else
+# define V8JS_TSRMLS_FETCH()
+#endif
+
+
 PHP_MINIT_FUNCTION(v8js_class);
 
 #endif /* V8JS_CLASS_H */
