+++ v8js_timer.h
@@ -14,6 +14,16 @@
 #ifndef V8JS_TIMER_H
 #define V8JS_TIMER_H
 
+// Timer context
+struct php_v8js_timer_ctx
+{
+  long time_limit;
+  long memory_limit;
+  std::chrono::time_point<std::chrono::high_resolution_clock> time_point;
+  php_v8js_ctx *v8js_ctx;
+  bool killed;
+};
+
 void v8js_timer_thread(TSRMLS_D);
 void v8js_timer_push(long time_limit, long memory_limit, php_v8js_ctx *c TSRMLS_DC);
