+++ v8js_v8.h
@@ -14,6 +14,33 @@
 #ifndef V8JS_V8_H
 #define V8JS_V8_H
 
+/* Helper macros */
+#define V8JS_SYM(v)			v8::String::NewFromUtf8(isolate, v, v8::String::kInternalizedString, sizeof(v) - 1)
+#define V8JS_SYML(v, l)		v8::String::NewFromUtf8(isolate, v, v8::String::kInternalizedString, l)
+#define V8JS_STR(v)			v8::String::NewFromUtf8(isolate, v)
+#define V8JS_STRL(v, l)		v8::String::NewFromUtf8(isolate, v, v8::String::kNormalString, l)
+#define V8JS_INT(v)			v8::Integer::New(isolate, v)
+#define V8JS_UINT(v)		v8::Integer::NewFromUnsigned(isolate, v)
+#define V8JS_FLOAT(v)		v8::Number::New(isolate, v)
+#define V8JS_BOOL(v)		((v)?v8::True(isolate):v8::False(isolate))
+#define V8JS_DATE(v)		v8::Date::New(isolate, v)
+#define V8JS_NULL			v8::Null(isolate)
+#define V8JS_UNDEFINED		v8::Undefined(isolate)
+#define V8JS_MN(name)		v8js_method_##name
+#define V8JS_METHOD(name)	void V8JS_MN(name)(const v8::FunctionCallbackInfo<v8::Value>& info)
+#define V8JS_THROW(isolate, type, message, message_len)	(isolate)->ThrowException(v8::Exception::type(V8JS_STRL(message, message_len)))
+#define V8JS_GLOBAL(isolate)			((isolate)->GetCurrentContext()->Global())
+
+
+/* Extracts a C string from a V8 Utf8Value. */
+static inline const char * ToCString(const v8::String::Utf8Value &value) /* {{{ */
+{
+	return *value ? *value : "<string conversion failed>";
+}
+/* }}} */
+
+
+
 void v8js_v8_init(TSRMLS_D);
 void v8js_v8_call(php_v8js_ctx *c, zval **return_value,
 				  long flags, long time_limit, long memory_limit,
